Move PackedChineseBasedYearInfo into provider module, make ULE
